Add implementations for GetLog and change MessageHandler to ResponseHandler
